C++ Actors for movement.
Here are some C++ files I created to work as actors in Unreal Engine 5 that I made to apply movement to Blueprint Classes. I also took some screenshots of the Blueprint event graph used on some of the actors to show how their Boolean variables worked. The basic movement for the classes was taught in the Udemy course, however, I implemented the launch mechanic that would work with the rotation mechanic to act as a launch pad. I also created a function that moves the platform back to its original location after the desired rotation has been reached. Lastly, I created a separate class called Elevator that moves the platform any time the player triggers a event begin overlap and moves back to its original location when the event end overlap is triggered.